Index.html changes
Began actual content.
Changed it so links are yellow, nav bar is below header and to the left.
For now I plan to just have it floating up there.
I plan to try to include all 4 of the sections in one div and scroll to
them once I learn to use the "scrollTo".